finished testing, working on readme

+++ main/csci2720/p1/LinkedListDriver.java
@@ -9,6 +9,21 @@ import java.io.FileNotFoundException;
  */
 public class LinkedListDriver {
 
+    /*
+     * Constant commands
+     */
+    static final char INSERT = 'i';
+    static final char DELETE = 'd';
+    static final char SEARCH = 's';
+    static final char GET_NEXT = 'n';
+    static final char RESET_LIST = 'r';
+    static final char DEL_ALT = 'a';
+    static final char MERGE = 'm';
+    static final char INTERSECTION = 't';
+    static final char PRINT_ALL = 'p';
+    static final char LENGTH = 'l';
+    static final char QUIT = 'q';
+
     /**
      * The main method.
      * @param args the command line arguments
@@ -46,14 +61,15 @@ public class LinkedListDriver {
         System.out.print("Enter a command: ");
         String command = stdIn.nextLine();
         ItemType itemtype = new ItemType(0);
-        while (!command.equals("q")) {
+        if (command.equals("" + QUIT)) System.out.println("Exiting the program...");
+        while (!command.equals("" + QUIT)) {
             boolean valid = true;
             switch (command) {
-            case "p": // print list
+            case "" + PRINT_ALL: // print list
                 System.out.print("The list is: ");
                 list.printList();
                 break;
-            case "i": // insert item
+            case "" + INSERT: // insert item
                 System.out.print("Enter a number to insert: ");
                 int item = Integer.parseInt(stdIn.nextLine());
                 itemtype = new ItemType(item);
@@ -67,44 +83,50 @@ public class LinkedListDriver {
                 System.out.print("New list: ");
                 list.printList();
                 break;
-            case "d": // delete item
+            case "" + DELETE: // delete item
                 System.out.print("Enter a number to delete: ");
                 itemtype = new ItemType(Integer.parseInt(stdIn.nextLine()));
-                System.out.print("Original list: ");
-                list.printList();
-                list.deleteItem(itemtype);
-                System.out.print("New list: ");
-                list.printList();
+                if (list.getLength() == 0) {
+                    System.out.println("You cannot delete from an empty list");
+                } else {
+                    System.out.print("Original list: ");
+                    list.printList();
+                    list.deleteItem(itemtype);
+                    System.out.print("New list: ");
+                    list.printList();
+                }
                 break;
-            case "l": // print length of list
+            case "" + LENGTH: // print length of list
                 System.out.print("The length of the list is " + list.getLength() + "\n");
                 break;
-            case "s": // search item
+            case "" + SEARCH: // search item
                 System.out.print("Enter a number to search: ");
                 itemtype = new ItemType(Integer.parseInt(stdIn.nextLine()));
                 System.out.print("Original list: ");
                 list.printList();
-                if (list.searchItem(itemtype) == -1) {
+                if (list.getLength() == 0) {
+                    System.out.println("The list is empty");
+                } else if (list.searchItem(itemtype) == -1) {
                     System.out.println("Item is not in the list");
                 } else {
                     System.out.println("The item is present at index " + list.searchItem(itemtype));
                 } // else
                 break;
-            case "n": // get next item
+            case "" + GET_NEXT: // get next item
                 list.getNextItem();
                 break;
-            case "r": // reset iterator
+            case "" + RESET_LIST: // reset iterator
                 list.resetList();
                 System.out.println("Iterator is reset");
                 break;
-            case "a": // delete alternate nodes
+            case "" + DEL_ALT: // delete alternate nodes
                 System.out.print("Original list: ");
                 list.printList();
                 tempList = list.deleteAltNodes(list);
                 System.out.print("Modified list: ");
                 tempList.printList();
                 break;
-            case "m": // merge list
+            case "" + MERGE: // merge list
                 System.out.print("Enter the length of the new list: ");
                 int length = stdIn.nextInt();
                 System.out.print("Enter the numbers: ");
@@ -118,12 +140,12 @@ public class LinkedListDriver {
                 list.printList();
                 System.out.print("The list 2: ");
                 list2.printList();
-                list = list.mergeList(list, list2);
+                list2 = list.mergeList(list, list2);
                 System.out.print("Merged list: ");
-                list.printList();
+                list2.printList();
                 stdIn.nextLine();
                 break;
-            case "t": // intersection
+            case "" + INTERSECTION: // intersection
                 System.out.print("Enter the length of the new list: ");
                 int listLength = stdIn.nextInt();
                 System.out.print("Enter the numbers: ");
@@ -149,6 +171,7 @@ public class LinkedListDriver {
             if (valid) System.out.print("Enter a command: ");
             else System.out.print("Invalid command try again: ");
             command = stdIn.nextLine();
+            if (command.equals("" + QUIT)) System.out.println("Exiting the program...");
         } // while
     } // main
 } // class
